// OmkarFCUBEAPP/src/app/models/brsentrymodel.ts
export class Brsentrymodel {
   
    transId: string = "";
    transDate: string = "";
    cardNo: string = "";
    bankAc: string = "";
    docNo: string = "";
    debitRs: string = "";
    creditRs: string = "";
    chequeNo: string = "";
    chequeDate: string = "";
    narration: string = "";
    clearDate: string = "";
    loggedInUser: string = "";

   // clearDate: string = "";
    
}